Refactor notes
